Added separate umount shutdown service to fix "halt".
   Separate service and a "shutdown" parameter introduced
   for umount script. (HYP-2816)

+++ meta-hidav/recipes-hidav/rootfs-overlay/rootfs-overlay/rootfs-overlay-sources/umount-rootfs-overlay.sh
@@ -77,17 +77,16 @@ killprocs=`fuser -m ${pivot_root_mountpoint}`
 umount -l ${pivot_root_mountpoint}
 umount -l ${overlays_data_mountpoint}
 
-# if this script has been started from the overlay root
-# in an interactive shell then this will most probably 
-# kill the script itself
-exec chroot . sh -c "
-    cd /
-    trap \"\" SIGTERM
-    kill -TERM ${killprocs}
-    sleep 1
-    kill -KILL ${killprocs}
-" <dev/console >dev/console 2>&1
-
-
-
+if [ "$1" != "shutdown" ] ; then
+    # if this script has been started from the overlay root
+    # in an interactive shell then this will most probably 
+    # kill the script itself.
+    exec chroot . sh -c "
+        cd /
+        trap \"\" SIGTERM
+        kill -TERM ${killprocs}
+        sleep 1
+        kill -KILL ${killprocs}
+    " <dev/console >dev/console 2>&1
+fi
